// src/activity/enums/timeline-event-type.enum.ts
export enum TimelineEventType {
  APPLICATION_SUBMITTED = 'APPLICATION_SUBMITTED',
  REVIEW_STARTED = 'REVIEW_STARTED',
  CORRECTION_REQUIRED = 'CORRECTION_REQUIRED',
  CORRECTION_RESUBMITTED = 'CORRECTION_RESUBMITTED',
  DOCUMENTS_APPROVED = 'DOCUMENTS_APPROVED',
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  PAYMENT_CONFIRMED = 'PAYMENT_CONFIRMED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  APPOINTMENT_SCHEDULED = 'APPOINTMENT_SCHEDULED',
  APPOINTMENT_CANCELLED = 'APPOINTMENT_CANCELLED',
  APPOINTMENT_NO_SHOW = 'APPOINTMENT_NO_SHOW',
  READY_FOR_INSPECTION = 'READY_FOR_INSPECTION',
  INSPECTION_PASSED = 'INSPECTION_PASSED',
  INSPECTION_FAILED = 'INSPECTION_FAILED',
  STICKER_PREPARING = 'STICKER_PREPARING',
  STICKER_READY = 'STICKER_READY',
  STICKER_ISSUED = 'STICKER_ISSUED',
  APPLICATION_COMPLETED = 'APPLICATION_COMPLETED',
  APPLICATION_CANCELLED = 'APPLICATION_CANCELLED',
}